Implementation Feature-slace design architecture
