+++ src/app/types/Global.d.ts
