+++ src/shared/context/index.ts
@@ -1,0 +1,4 @@
+import { ThemeProvider } from "./theme-context/ThemeContext";
+import { useTheme } from "./theme-context/useTheme";
+
+export { ThemeProvider, useTheme };
