+++ src/shared/context/theme-context/ThemeContext.tsx
@@ -1,0 +1,42 @@
+import { FC, useMemo, useState, ReactNode, createContext } from "react";
+
+export enum Theme {
+  LIGHT = "light",
+  DARK = "dark",
+}
+
+export interface ThemContextProps {
+  theme?: Theme;
+  setTheme?: (theme: Theme) => void;
+}
+
+export const ThemeContext = createContext<ThemContextProps>({});
+
+export const LOCAL_STORAGE_THEM_KEY = "theme";
+
+const defaultTheme =
+  (localStorage.getItem(LOCAL_STORAGE_THEM_KEY) as Theme) || Theme.LIGHT;
+
+type ThemeProviderProps = {
+  children: ReactNode;
+};
+
+export const ThemeProvider: FC<ThemeProviderProps> = ({ children }) => {
+  const [theme, setTheme] = useState<Theme>(defaultTheme);
+
+  const defaultProps = useMemo(
+    () => ({
+      theme: theme,
+      setTheme: setTheme,
+    }),
+    [theme]
+  );
+
+  return (
+    <ThemeContext.Provider value={defaultProps}>
+      {children}
+    </ThemeContext.Provider>
+  );
+};
+
+// export default ThemeProvider;
