--- src/theme/ThemeProvider.tsx
@@ -1,29 +1,0 @@
-import { FC, useMemo, useState, ReactNode } from "react";
-import { LOCAL_STORAGE_THEM_KEY, Theme, ThemeContext } from "./ThemeContext";
-
-const defaultTheme =
-  (localStorage.getItem(LOCAL_STORAGE_THEM_KEY) as Theme) || Theme.LIGHT;
-
-type ThemeProviderProps = {
-  children: ReactNode;
-};
-
-const ThemeProvider: FC<ThemeProviderProps> = ({ children }) => {
-  const [theme, setTheme] = useState<Theme>(defaultTheme);
-
-  const defaultProps = useMemo(
-    () => ({
-      theme: theme,
-      setTheme: setTheme,
-    }),
-    [theme]
-  );
-
-  return (
-    <ThemeContext.Provider value={defaultProps}>
-      {children}
-    </ThemeContext.Provider>
-  );
-};
-
-export default ThemeProvider;
